backend/pyproject.toml
[tool.poetry]
authors = ["admin <admin@example.com>"]
description = "FastAPI project generated using minimal-fastapi-postgres-template."
name = "backend"
version = "0.1.0-alpha"

[tool.poetry.dependencies]
python = "^3.12"

alembic = "^1.12.1"
asyncpg = "^0.29.0"
fastapi = "^0.104.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = { extras = ["dotenv", "email"], version = "^2.4.2" }
pydantic-settings = "^2.0.3"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
coverage = "^7.3.2"
httpx = "^0.25.1"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.4"
uvicorn = { extras = ["standard"], version = "^0.24.0" }


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "-v"
asyncio_mode = "auto"
filterwarnings = [
    # Passlib 1.7.4 depends on crypt - https://foss.heptapod.net/python-libs/passlib/-/issues/148
    "ignore:'crypt' is deprecated and slated for removal",
]
markers = ["pytest.mark.asyncio"]
minversion = "6.0"
testpaths = ["app/tests"]

[tool.ruff]
target-version = "py312"
# pycodestyle, pyflakes, isort, pylint, pyupgrade
select = ["E", "W", "F", "I", "PL", "UP"]
ignore = ["E501"]
